Remove mergify (using pull, see below)
Mergify was used when the repo still used Backstroke to auto-pull from upstream, but Backstroke shut down, so I switched to using pull[bot] which has the feature of auto-merging as well, and is safer because it only applies to the PR's that it pulls from upstream. Mergify is not needed anymore, though it would "beat" pull to the task of auto-merging by a second, which is a toss up and why the last few commits were merged by Mergify. 🤷